pyproject.toml
[project]
name = "multi-agent-quant"
version = "0.1.0"
description = "Advanced multi-agent quantitative trading framework"
authors = [
  { name = "AI Task Codex" }
]
requires-python = ">=3.10"
dependencies = [
  "pyyaml",
  "pydantic>=2.6",
  "rich",
  "tushare",
]
keywords = ["multi-agent", "quant", "trading", "llm", "evolution"]
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://example.com/multi-agent-quant"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
